requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
docopt==0.6.2
funcsigs==1.0.2
functools32==3.2.3.post2
h5py==2.7.0
html5lib==0.9999999
jsonpickle==0.9.5
Markdown==2.6.8
matplotlib==2.0.2
mock==2.0.0
munch==2.2.0
networkx==1.11
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pbr==3.1.1
Pillow==4.2.1
protobuf==3.3.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
sacred==0.7.0
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.2.1
Werkzeug==0.12.2
wrapt==1.10.10
